+++ codegen/gql_code_builder/lib/ast.dart
@@ -10,14 +10,14 @@ Library buildAstLibrary(
   SourceNode source,
 ) {
   final definitions = source.document.definitions.map(
-    (def) => fromNode(def).assignConst(identifier(_getName(def))).statement,
+    (def) => declareConst(identifier(_getName(def))).assign(fromNode(def)).statement,
   );
 
-  final document = refer(
-    "DocumentNode",
-    "package:gql/ast.dart",
-  )
-      .call(
+  final document = declareConst("document")
+      .assign(refer(
+        "DocumentNode",
+        "package:gql/ast.dart",
+      ).call(
         [],
         {
           "definitions": literalList(
@@ -29,8 +29,7 @@ Library buildAstLibrary(
                 ),
           ),
         },
-      )
-      .assignConst("document")
+      ))
       .statement;
 
   return Library(