+++ examples/gql_example_flutter/lib/src/all_pokemon/all_pokemon.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:gql_exec/gql_exec.dart';
 
-import './graphql/all_pokemon.data.gql.dart';
-import './graphql/all_pokemon.req.gql.dart';
+import './graphql/__generated__/all_pokemon.data.gql.dart';
+import './graphql/__generated__/all_pokemon.req.gql.dart';
 import '../config.dart';
 import '../pokemon_card/pokemon_card.dart';
 
@@ -30,10 +30,13 @@ class AllPokemonScreen extends StatelessWidget {
 
           final data = GAllPokemonData.fromJson(snapshot.data!.data!);
 
+          final pokemons =
+              data!.pokemons!.whereType<GAllPokemonData_pokemons>().toList();
+
           return ListView.builder(
-            itemCount: data!.pokemons!.length,
+            itemCount: pokemons.length,
             itemBuilder: (context, index) => PokemonCard(
-              pokemon: data.pokemons![index],
+              pokemon: pokemons[index],
             ),
           );
         },